Move chat and vote pages into game folder
