+++ client/src/app/app.module.ts
@@ -12,8 +12,8 @@ import { StartGameOptionButtonsComponent } from './start-game-option-buttons/sta
 import { EnterGameCodeScreenComponent } from './enter-game-code-screen/enter-game-code-screen.component';
 import { NewGameCodeScreenComponent } from './new-game-code-screen/new-game-code-screen.component';
 import { LobbyComponent } from './lobby/lobby.component';
-import { ChatComponent } from './chat/chat.component';
-import { VotingPageComponent } from './voting-page/voting-page.component';
+import { ChatComponent } from './game-page/chat/chat.component';
+import { VotingPageComponent } from './game-page/voting-page/voting-page.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 @NgModule({
