+++ client/src/app/game-page/voting-page/progress-spinner-configurable-example/progress-spinner-configurable-example.component.html
