+++ client/src/app/game-page/voting-page/radio-ng-model-example/radio-ng-model-example.component.ts
