+++ client/src/app/game-page/voting-page/voting-page.component.ts
